+++ tests/test_master_acks.py
@@ -2,10 +2,13 @@
 import xmostest
 from i2c_master_checker import I2CMasterChecker
 
-def runtest():
+def do_master_test(arch, stop):
     resources = xmostest.request_resource("xsim")
 
-    binary = 'i2c_master_test/bin/tx_only/i2c_master_test_tx_only.xe'
+    binary = 'i2c_master_test/bin/tx_only_%(stop)s_%(arch)s/i2c_master_test_tx_only_%(stop)s_%(arch)s.xe' % {
+      'stop' : stop,
+      'arch' : arch,
+    }
 
     checker = I2CMasterChecker("tile[0]:XS1_PORT_1A",
                                "tile[0]:XS1_PORT_1B",
@@ -15,9 +18,10 @@ def runtest():
                                              True, True, False,
                                              False, True])
 
-    tester = xmostest.ComparisonTester(open('ack_test.expect'),
+    tester = xmostest.ComparisonTester(open('ack_test_%s.expect' % stop),
                                       'lib_i2c', 'i2c_master_sim_tests',
-                                      'ack_test', {'speed':400},
+                                      'ack_test',
+                                      {'speed':400, 'arch' : arch, 'stop' : stop},
                                       regexp=True)
 
     xmostest.run_on_simulator(resources['xsim'], binary,
@@ -25,3 +29,8 @@ def runtest():
                               simargs=['--weak-external-drive'],
                               suppress_multidrive_messages = True,
                               tester = tester)
+
+def runtest():
+  for arch in ['xs1', 'xs2']:
+    for stop in ['stop', 'no_stop']:
+        do_master_test(arch, stop)